// src/types/enum.d.ts
export declare enum DateFormat {
    WITH_TIME = "YY.MM.DD HH:mm",
    YYMMDD = "YY.MM.DD",
    MMDD = "MM.DD"
}
export declare enum Language {
    KOREAN = "KOREAN",
    ENGLISH = "ENGLISH",
    JAPANESE = "JAPANESE",
    CHINESE = "CHINESE"
}
export declare enum WindowSizeNumber {
    MOBILE = 400,
    PHABLET = 550,
    TABLET = 750,
    DESKTOP = 1000,
    DESKTOPHD = 1200
}
export declare enum WindowSizeShort {
    MOBILE = "sm",
    PHABLET = "md",
    TABLET = "wmd",
    DESKTOP = "lg",
    DESKTOPHD = "wlg"
}
export declare enum WindowSize {
    sm = "sm",
    md = "sm",
    wmd = "wmd",
    lg = "lg",
    wlg = "wlg",
    full = "full"
}
export declare type TNationalShort = "kr" | "gb" | "jp" | "cn";
export declare type TLanguageShort = "kr" | "en" | "jp" | "cn";
export declare enum LanguageShortResverse {
    kr = "KOREAN",
    jp = "JAPANESE",
    cn = "CHINESE",
    en = "ENGLISH"
}
export declare enum LanguageResverseShort {
    KOREAN = "kr",
    ENGLISH = "en",
    JAPANESE = "jp",
    CHINESE = "cn"
}
export declare enum LangShortToNational {
    kr = "kr",
    en = "gb",
    jp = "jp",
    cn = "cn"
}
export declare enum LanguageItSelf {
    KOREAN = "\uD55C\uAD6D\uC5B4",
    ENGLISH = "English",
    JAPANESE = "\u65E5\u672C\u8A9E",
    CHINESE = "\u4E2D\u570B\u8A9E"
}
export declare type TMarginSize = "no" | "largest" | "huge" | "large" | "normal" | "small" | "tiny" | "superTiny";
export declare type TextAlign = "left" | "right" | "center";
export declare type TextSize = "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "large" | "normal" | "small" | "tiny" | "superTiny";
export declare enum RoomGender {
    FEMALE = "FEMALE",
    MALE = "MALE",
    ANY = "ANY",
    SEPARATELY = "SEPARATELY"
}
export declare enum TimePerMs {
    DAY = 86400000,
    H = 3600000,
    M = 60000
}
export declare enum Gender {
    FEMALE = "FEMALE",
    MALE = "MALE"
}
export declare enum ExcelExpress {
    SELECT_OP = "SELECT_OP",
    DATE_OP = "DATE_OP",
    COUNT_OP = "COUNT_OP"
}
export declare enum NotiType {
    ELSE = "ELSE",
    NEW_BOOKING = "NEW_BOOKING",
    PRODUCT_EXPIRE = "PRODUCT_EXPIRE",
    TO_ALL = "TO_ALL"
}
export declare type IconSize = "tiny" | "small" | "normal" | "large" | "huge" | "largest" | "largest2";
export declare type JDColor = "darkPrimary" | "primary" | "third" | "point" | "new" | "warn" | "error" | "black" | "white" | "grey1" | "grey2" | "grey3" | "grey4" | "grey5" | "normal" | "blue" | "positive";
export declare enum Funnels {
    AGODA = "AGODA",
    AIRBNB = "AIRBNB",
    BOOKING_COM = "BOOKING_COM",
    COOPANG = "COOPANG",
    ELSE_CHANNEL = "ELSE_CHANNEL",
    FREINDS = "FREINDS",
    HOMEPAGE = "HOMEPAGE",
    NAVER = "NAVER",
    PHONE_CALL = "PHONE_CALL",
    WALK_IN = "WALK_IN",
    YANOLJA = "YANOLJA",
    YEOGIEOTTAE = "YEOGIEOTTAE"
}
export declare enum SendTarget {
    BOTH = "BOTH",
    GUEST = "GUEST",
    HOST = "HOST"
}
export declare enum LayoutType {
    Layout_A = "Layout_A",
    Layout_B = "Layout_B"
}
export declare enum Day {
    FRI = "FRI",
    MON = "MON",
    SAT = "SAT",
    SUN = "SUN",
    THU = "THU",
    TUE = "TUE",
    WED = "WED"
}